Make Invoke-SqlCommand async.
This fixes an NRE during handling of a server message when output is
redirected to a file.  The NRE occurred because when using SqlClient
asynchronously, server message events can fire on any thread, but only
the main thread can write cmdlet output.  The code before this commit
did not make any attempt to marshal the output back to the main thread.